requirements.txt
aiogram
openai
pydantic>=2.0
python-dotenv
aiohttp
flask
asyncpg



